refactor: apply feature first strategy
