+++ qaul_ui/lib/screens/home/home_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../decorators/qaul_nav_bar_decorator.dart';
+import '../../nav_bar/widgets/qaul_nav_bar_decorator.dart';
 import '../../providers/providers.dart';
 import 'dynamic_network/dynamic_network_screen.dart';
 import 'tabs/tab.dart';
