+++ qaul_ui/test/decorators/qaul_nav_bar_item_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:qaul_rpc/qaul_rpc.dart';
-import 'package:qaul_ui/decorators/qaul_nav_bar_decorator.dart';
+import 'package:qaul_ui/nav_bar/widgets/qaul_nav_bar_decorator.dart';
 import 'package:qaul_ui/l10n/app_localizations.dart';
 import 'package:qaul_ui/providers/providers.dart';
 import 'package:qaul_ui/qaul_app.dart';
